// src/swagger/model/complianceReport.ts
/**
 * fms
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ComplianceReport { 
    zoneName?: string;
    regionName?: string;
    fitnessNotUploaded?: number;
    fitnessInvalid?: number;
    fitnessLegal?: number;
    fitnessScrap?: number;
    fitnessValid?: number;
    fitnessTotal?: number;
    fitnessCompliant?: number;
    roadTaxNotUploaded?: number;
    roadTaxInvalid?: number;
    roadTaxLegal?: number;
    roadTaxScrap?: number;
    roadTaxValid?: number;
    roadTaxTotal?: number;
    roadTaxCompliant?: number;
}